Extract makeApiCalls from LMSF to VM
